# perf: improve applyOrdering by avoid call to GetByCurrentId