@chlunde 2024-01-19
> @chlunde thanks a lot for your perf PRs, they are greatly needed and will massively improve the user experience. I left 2 comments for small changes but overall if this passes tests it should be good to go.

@yanniszark I believe I've addressed your comments, could we get this merged now?