@koba1t 2024-02-06
HI @yanniszark 
If you think your commented point was fixed, Please type the `/lgtm` command.
https://github.com/kubernetes-sigs/kustomize/pull/5079#pullrequestreview-1336074940
That mark means you think this PR is ready to merge.